+++ src/infrastructure/adapters/ConsoleEffectSink.js
@@ -11,7 +11,7 @@ import { createDeliveryObservation } from '../../domain/types/DeliveryObservatio
 
 /**
  * @typedef {import('../../domain/types/EffectEmission.js').EffectEmission} EffectEmission
- * @typedef {import('../../domain/types/DeliveryLens.js').DeliveryLens} DeliveryLens
+ * @typedef {import('../../domain/types/ExternalizationPolicy.js').ExternalizationPolicy} ExternalizationPolicy
  */
 
 export class ConsoleEffectSink extends EffectSinkPort {
@@ -31,7 +31,7 @@ export class ConsoleEffectSink extends EffectSinkPort {
 
   /**
    * @param {EffectEmission} emission
-   * @param {DeliveryLens} lens
+   * @param {ExternalizationPolicy} lens
    * @returns {Promise<import('../../domain/types/DeliveryObservation.js').DeliveryObservation>}
    */
   deliver(emission, lens) {